Rest of the tests added
