+++ src/components/Triangle/__tests__/Triangle.test.tsx
@@ -1,23 +1,23 @@
 import React from 'react';
 import { act, fireEvent, render, waitFor } from '@testing-library/react-native';
-import Hexagon from '../';
-import { hexagon, hexagonForTestingInfo } from '../../../mocks/hexagon';
+import Triangle from '../';
+import { triangle, triangleForTestingInfo } from '../../../mocks/triangle';
 import { IChartProps } from '../../../types';
 
 test('renders correctly', async () => {
   const rendered = render(
-    <Hexagon {...hexagon} style={{width: 300, height: 300}} />,
+    <Triangle {...triangle} style={{width: 300, height: 300}} />,
   );
 
   expect(rendered).toBeTruthy();
 });
 
 test('texts render and press correctly', async () => {
-  const functions = [...new Array(6)].map(() => jest.fn());
+  const functions = [...new Array(3)].map(() => jest.fn());
 
   const data: IChartProps = {
-    ...hexagonForTestingInfo,
-    poles: hexagonForTestingInfo.poles.map((v, i) => ({
+    ...triangleForTestingInfo,
+    poles: triangleForTestingInfo.poles.map((v, i) => ({
       ...v,
       info: {...v.info, onPress: functions[i]},
     })),
@@ -25,7 +25,7 @@ test('texts render and press correctly', async () => {
 
   jest.useFakeTimers();
   const {getByTestId, getByText} = render(
-    <Hexagon
+    <Triangle
       {...data}
       animation={undefined}
       style={{width: 300, height: 300}}
